Move `tests/unit/` to `tests/slice/`
