+++ ide/src/tests/slice/index.ts
